+++ dms-20250414/ChangeLog.md
@@ -1,3 +1,18 @@
+2026-02-10 Version: 1.8.7
+- Update API CreateDataAgentSession: add request parameters SessionConfig.EncryptKey.
+- Update API CreateDataAgentSession: add request parameters SessionConfig.EncryptType.
+- Update API CreateDataAgentSession: add request parameters SessionConfig.ReportPageWidth.
+- Update API CreateDataAgentSession: add request parameters SessionConfig.ReportWaterMark.
+- Update API CreateDataAgentSession: add response parameters Body.Data.SessionConfig.EncryptKey.
+- Update API CreateDataAgentSession: add response parameters Body.Data.SessionConfig.EncryptType.
+- Update API CreateDataAgentSession: add response parameters Body.Data.SessionConfig.ReportPageWidth.
+- Update API CreateDataAgentSession: add response parameters Body.Data.SessionConfig.ReportWaterMark.
+- Update API DescribeDataAgentSession: add response parameters Body.Data.SessionConfig.EncryptKey.
+- Update API DescribeDataAgentSession: add response parameters Body.Data.SessionConfig.EncryptType.
+- Update API DescribeDataAgentSession: add response parameters Body.Data.SessionConfig.ReportPageWidth.
+- Update API DescribeDataAgentSession: add response parameters Body.Data.SessionConfig.ReportWaterMark.
+
+
 2026-02-03 Version: 1.8.6
 - Update API SendChatMessage: add request parameters ParentSessionId.